Implement element geometry mapping as identity
In absense of a reference element for geometry type `none` we allow the
grid to choose its own reference element (on a per element basis). The
only obvious restriction is the world dimension, which must coincide
with the element dimension. For PolygonGrid elements we may thus simply
choose the identity mapping. For lines and vertices, we use the linear
mapping.